Update for issue #7 -Add Composer Scene Management -Add Folder Structure for scenes and functions -Move functions to helper file -Add property to access tile instance from group level
